MVC v3.0  Complite all CRUD-services and fixed bags

+++ API/application/controllers/MainCourseController.php
@@ -12,7 +12,7 @@ class MainCourseController extends Controller
 
         $model = new MainCourse();
         
-        if(array_key_exists('index', $_POST)) {
+        if((array_key_exists('index', $_POST)) && (!empty($_POST["index"]))) {
             $index = $_POST['index'];
             $model->deleteRecord($index);
         }
