+++ API/application/controllers/MainProfessorController.php
@@ -12,9 +12,13 @@ class MainProfessorController extends Controller
 
         $model = new MainProfessor();
         
-        if(array_key_exists('index', $_POST)) {
+        if((array_key_exists('index', $_POST)) && (!empty($_POST["index"]))) {
             $index = $_POST['index'];
             $model->deleteRecord($index);
+        }
+
+        if(array_key_exists('id',$_POST)) {
+            $model->updateRecord($_POST);
         }
 
         $data = $model->getDataFormTable();
