+++ API/application/controllers/MainStudentsOfCoursesController.php
@@ -12,9 +12,13 @@ class MainStudentsOfCoursesController extends Controller
 
         $model = new MainStudentsOfCourses();
         
-        if(array_key_exists('index', $_POST)) {
+        if((array_key_exists('index', $_POST)) && (!empty($_POST["index"]))) {
             $index = $_POST['index'];
             $model->deleteRecord($index);
+        }
+
+        if(array_key_exists('idStudent',$_POST)) {
+            $model->updateRecord($_POST);
         }
 
         $data = $model->getDataFormTable();
